9/FactChecker/requirements.txt
semantic-text-splitter
python-dotenv
qdrant-client
brave-search
docling
chonkie
joblib
agno